+++ R/init_datasets.R
@@ -35,9 +35,9 @@ set_datasets_data <- function(datasets, data) {
   }
 
   for (idx in seq_along(data)) {
-    datasets$set_data(data[[idx]][["dataname"]], data[[idx]][["data"]])
+    datasets$set_data(data[[idx]][["dataname"]], data[[idx]]$raw_data)
     datasets$set_data_attr(data[[idx]][["dataname"]], "keys", data[[idx]][["keys"]])
-    datasets$set_data_attr(data[[idx]][["dataname"]], "column_labels", data[[idx]][["column_labels"]])
+    datasets$set_data_attr(data[[idx]][["dataname"]], "column_labels", data[[idx]]$get_column_labels())
     datasets$set_data_attr(data[[idx]][["dataname"]], "data_label", data[[idx]][["data_label"]])
   }